+++ test/functional/group-blocks.js
@@ -1,0 +1,104 @@
+/* eslint-env node, mocha */
+/* global $pg_database */
+import expect from 'unexpected';
+
+import cleanDB from '../dbCleaner';
+
+import {
+  authHeaders,
+  createGroupAsync,
+  createTestUsers,
+  performJSONRequest,
+  promoteToAdmin,
+  subscribeToAsync,
+} from './functional_test_helper';
+
+describe('Group Blocks', () => {
+  beforeEach(() => cleanDB($pg_database));
+
+  let luna, mars, venus, jupiter, selenites;
+
+  beforeEach(async () => {
+    [luna, mars, venus, jupiter] = await createTestUsers(['luna', 'mars', ' venus', 'jupiter']);
+    selenites = await createGroupAsync(luna, 'selenites');
+    await Promise.all([promoteToAdmin(selenites, luna, venus), subscribeToAsync(mars, selenites)]);
+  });
+
+  it(`should block Mars in Selenites`, async () => {
+    {
+      const resp = await blockUserInGroup(mars, selenites, luna);
+      expect(resp, 'to satisfy', { __httpCode: 200, blockedUsers: [mars.user.id] });
+    }
+
+    {
+      const resp = await getBlockedUsers(selenites, luna);
+      expect(resp, 'to satisfy', { __httpCode: 200, blockedUsers: [mars.user.id] });
+    }
+  });
+
+  it(`should not block Venus (as admin) in Selenites`, async () => {
+    const resp = await blockUserInGroup(venus, selenites, luna);
+    expect(resp, 'to satisfy', { __httpCode: 403 });
+  });
+
+  it(`should not allow Mars (as non-admin) to block anyone in Selenites`, async () => {
+    const resp = await blockUserInGroup(jupiter, selenites, mars);
+    expect(resp, 'to satisfy', { __httpCode: 403 });
+  });
+
+  describe('Mars is blocked in Selenites', () => {
+    beforeEach(() => blockUserInGroup(mars, selenites, luna));
+
+    it(`should allow Luna to unblock Mars`, async () => {
+      const resp = await unblockUserInGroup(mars, selenites, luna);
+      expect(resp, 'to satisfy', { __httpCode: 200, blockedUsers: [] });
+    });
+
+    it(`should not allow Jupiter (as non-admin) to unblock Mars`, async () => {
+      const resp = await unblockUserInGroup(mars, selenites, jupiter);
+      expect(resp, 'to satisfy', { __httpCode: 403 });
+    });
+
+    it(`should not allow Mars to create post in Selenites`, async () => {
+      const resp = await performJSONRequest(
+        'POST',
+        '/v1/posts',
+        {
+          post: { body: 'Hello' },
+          meta: { feeds: [selenites.username] },
+        },
+        authHeaders(mars),
+      );
+      expect(resp, 'to satisfy', { __httpCode: 403 });
+    });
+  });
+});
+
+// Helpers
+
+function blockUserInGroup(user, group, admin) {
+  return performJSONRequest(
+    'POST',
+    `/v2/groups/${group.username}/block/${user.username}`,
+    {},
+    authHeaders(admin),
+  );
+}
+
+function unblockUserInGroup(user, group, admin) {
+  return performJSONRequest(
+    'POST',
+    `/v2/groups/${group.username}/unblock/${user.username}`,
+    {},
+    authHeaders(admin),
+  );
+}
+
+function getBlockedUsers(group, admin) {
+  return performJSONRequest(
+    'GET',
+    `/v2/groups/${group.username}/blockedUsers`,
+    {},
+    authHeaders(admin),
+  );
+}